Preparations for version 1.2... - Made most dialogs window modal (except about, preferences and check for updates) instead of application modal, and are also now sheets on Mac OS X - Open and save file dialogs are now window modal and are sheets on Mac OS X - All message boxes are now window modal and are sheets on Mac OS X - Added menu items to the dock menu on Mac OS X - Moved "Lock Database" from the file menu to the application menu on Mac OS X - Database creation wizard now sets input focus and default buttons correctly - Improved the size of windows and dialogs in the application - Fixed a bug in the preferences dialog preventing the restore defaults button from working - Improved the open database dialog - Added a native search bar to the toolbar on Mac OS X - Added support for "run at login" in the preferences for Mac OS X - Added drag and drop support for opening database files - Printing has been fixed for all platforms - previously it would begin printing before the user selected "print" or "cancel" on the print dialog - Fullscreen mode temporarily removed due to Qt issues - Very early stages of Android support
Apparently there is a bug where the About menu disappears on Mac with more than one application window, and there is some sort of issue with inserting custom fields...

+++ src/3rdparty/configure.sh
@@ -1,18 +1,22 @@
 echo "This script configures 3rd party code for building with Petroules applications."
 echo "Copyright (c) 2010-2011 Petroules Corporation"
+echo ""
 
-# Detect operating system by first assuming we are 
-if [ $(uname -o) == "Msys" ]; then # we are Windows (MSYS)
-	BUILDDIR="temp-win32-gcc"
+# Set physical directory of shell script
+pushd $(dirname $0)
+
+# Detect operating system
+if [ $(uname) == "Linux" ]; then # we are some sort of Linux
+	# We only support 32-bit Linux at this point
+	BUILDDIR="$PWD/temp-linux32"
 fi
 
 if [ $(uname) == "Darwin" ]; then # we are Mac OS X
-	BUILDDIR="temp-mac64"
+	BUILDDIR="$PWD/temp-mac64"
 fi
 
-if [ $(uname) == "Linux" ]; then # we are some sort of Linux
-	# We only support 32-bit Linux at this point
-	BUILDDIR="temp-linux32"
+if [ $(uname -o) == "Msys" ]; then # we are Windows (MSYS)
+	BUILDDIR="$PWD/temp-win32-gcc"
 fi
 
 if [ $BUILDDIR == "" ]; then # unknown operating system, error
@@ -20,25 +24,33 @@ if [ $BUILDDIR == "" ]; then # unknown operating system, error
 	exit
 fi
 
-rm -rf "$BUILDDIR"
+# Clean previous make file and temp directory
+if [ -f "$PWD/make.sh" ]; then
+	rm "$PWD/make.sh"
+fi
+
+if [ -d "$BUILDDIR" ]; then
+	rm -rf "$BUILDDIR"
+fi
 
 if [ "$1" == "--clean" ]; then
+	echo "Removed temporary files"
 	exit
 fi
 
 mkdir "$BUILDDIR"
 
 # Set filenames
-BOTAN_FN=Botan-1.9.12
+BOTAN_FN=Botan-1.10.1
 QTSLNS_FN=qt-solutions-qt-solutions
 QTSLNS_AFN=$QTSLNS_FN-master
-LIEL_FN=liel-liel
-LIEL_AFN=$LIEL_FN-master
+SYNTEZA_FN=synteza-synteza
+SYNTEZA_AFN=$SYNTEZA_FN-master
 
 # Set directory locations
 ARCHIVES_DIR=$PWD/archives
 PETROULES_DIR=$PWD/petroules
-OUT_DIR=$PWD/$BUILDDIR
+OUT_DIR=$BUILDDIR
 
 echo Extracting Botan source...
 tar xzf $ARCHIVES_DIR/$BOTAN_FN.tgz $BOTAN_FN
@@ -48,14 +60,23 @@ echo Extracting Qt Solutions source...
 tar xzf $ARCHIVES_DIR/$QTSLNS_AFN.tar.gz $QTSLNS_FN
 mv $QTSLNS_FN $OUT_DIR/qtsolutions
 
-echo Extracting LIEL source...
-tar xzf $ARCHIVES_DIR/$LIEL_AFN.tar.gz $LIEL_FN
-mv $LIEL_FN $OUT_DIR/liel
+echo Extracting Synteza source...
+tar xzf $ARCHIVES_DIR/$SYNTEZA_AFN.tar.gz $SYNTEZA_FN
+mv $SYNTEZA_FN $OUT_DIR/synteza
 
-echo Extracting create-dmg tool...
-unzip $ARCHIVES_DIR/create-dmg.zip -d $OUT_DIR
+# Only need create-dmg for Mac OS X
+if [ $(uname) == "Darwin" ]; then
+	echo Extracting create-dmg tool...
+	unzip $ARCHIVES_DIR/create-dmg.zip -d $OUT_DIR
+fi
 
 echo Copying modified QtSingleApplication files...
 cp $PETROULES_DIR/qtsingleapplication.cpp $OUT_DIR/qtsolutions/qtsingleapplication/src
 cp $PETROULES_DIR/qtsingleapplication.h $OUT_DIR/qtsolutions/qtsingleapplication/src
 cp $PETROULES_DIR/botan.sh $OUT_DIR/botan
+cp $PETROULES_DIR/android.txt $OUT_DIR/botan/src/build-data/os
+
+echo Creating make script...
+echo $BUILDDIR/botan/botan.sh > "$PWD/make.sh"
+
+popd
